requirements.txt
streamlit==1.51.0
pandas==2.3.3
requests==2.32.5
python-dotenv==1.2.1
openai
tenacity
termcolor

